requirements.txt
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
docopt==0.6.2
eventlet==0.40.3
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.2
pythonanywhere==0.17.0
pythonanywhere-core==0.2.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
schema==0.7.8
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
snakesay==0.10.4
tabulate==0.9.0
typer==0.19.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
